+++ source/material/specular_reflective_material.cpp
@@ -2,16 +2,16 @@
 
 #include <cassert>
 
-SpecularReflectiveMaterial::SpecularReflectiveMaterial(float ior)
+SpecularReflectiveMaterial::SpecularReflectiveMaterial(double ior)
     : m_ior(ior)
 {
-    assert(m_ior >= 1.f);
+    assert(m_ior >= 1.0);
 }
 
 float3 SpecularReflectiveMaterial::bsdf(float3& ingoing, const float3& outgoing, const float2& random) const {
-    assert(equal(length(outgoing), 1.f));
-    assert(random[0] >= 0.f && random[0] < 1.f);
-    assert(random[1] >= 0.f && random[1] < 1.f);
+    assert(equal(length(outgoing), 1.0));
+    assert(random[0] >= 0.0 && random[0] < 1.0);
+    assert(random[1] >= 0.0 && random[1] < 1.0);
 
     if (outgoing.z < EPSILON) {
         return float3();
@@ -19,7 +19,7 @@ float3 SpecularReflectiveMaterial::bsdf(float3& ingoing, const float3& outgoing,
 
     ingoing = float3(-outgoing.x, -outgoing.y, outgoing.z);
 
-    return float3(fresnel_dielectric(ingoing.z, 1.f, m_ior) / std::abs(ingoing.z));
+    return float3(fresnel_dielectric(ingoing.z, 1.0, m_ior) / std::abs(ingoing.z));
 }
 
 bool SpecularReflectiveMaterial::is_specular() const {
